Rename provider / update README pre-reqs
This commit renames the provider from faas to openfaas, but may
need a change upstream in the faas-cli to work.

The test script now references an OpenFaaS Node.js template for
the hello-world function.

README updated with steps for pre-reqs.

The YAML file contains the address of the gateway URL - this is
needed when the URL is non-default i.e. 127.0.0.1:31112 for K8s.

Signed-off-by: Alex Ellis (VMware) <alexellis2@gmail.com>